refactor: use UserProfileStore in new header
