+++ packages/curve-common/src/widgets/Header/DesktopHeader.tsx
@@ -3,7 +3,7 @@ import Toolbar from '@mui/material/Toolbar'
 import Box from '@mui/material/Box'
 import Container from '@mui/material/Container'
 import { ConnectWalletIndicator } from 'curve-ui-kit/src/features/connect-wallet'
-import { UserProfileButton } from 'curve-ui-kit/src/features/user-profile'
+import { UserProfileButton, useUserProfileStore } from 'curve-ui-kit/src/features/user-profile'
 import { ChainSwitcher } from 'curve-ui-kit/src/features/switch-chain'
 import { AppButtonLinks } from './AppButtonLinks'
 import { HeaderLogo } from './HeaderLogo'
@@ -30,12 +30,16 @@ export const DesktopHeader = <TChainId extends number>({
   height, // height above + banner height
   pages,
   appStats,
-  themes: [theme, setTheme],
-  advancedMode,
   networkName,
   isLite = false,
 }: BaseHeaderProps<TChainId>) => {
   const [selectedApp, setSelectedApp] = useState<AppName>(currentApp)
+
+  const theme = useUserProfileStore((state) => state.theme)
+  const setTheme = useUserProfileStore((state) => state.setTheme)
+  const isAdvancedMode = useUserProfileStore((state) => state.isAdvancedMode)
+  const setAdvancedMode = useUserProfileStore((state) => state.setAdvancedMode)
+
   return (
     <>
       <AppBar color="transparent" ref={mainNavRef}>
@@ -56,7 +60,7 @@ export const DesktopHeader = <TChainId extends number>({
                 <UserProfileButton />
               ) : (
                 <>
-                  {advancedMode && <AdvancedModeSwitcher advancedMode={advancedMode} label={t`Advanced`} />}
+                  <AdvancedModeSwitcher advancedMode={[isAdvancedMode, setAdvancedMode]} label={t`Advanced`} />
                   <ThemeSwitcherButton theme={theme} onChange={setTheme} label={t`Mode`} />
                 </>
               )}
