+++ packages/curve-common/src/widgets/Header/types.ts
@@ -1,7 +1,6 @@
 import { ChainSwitcherProps } from 'curve-ui-kit/src/features/switch-chain'
 import { ConnectWalletIndicatorProps } from 'curve-ui-kit/src/features/connect-wallet'
 import { Dispatch, RefObject } from 'react'
-import type { ThemeKey } from 'curve-ui-kit/src/themes/basic-theme'
 import { AppName } from 'curve-ui-kit/src/shared/routes'
 import { GlobalBannerProps } from 'ui/src/Banner/GlobalBanner'
 
@@ -40,10 +39,7 @@ export type BaseHeaderProps<TChainId = number> = {
   height: string
   pages: AppPage[]
   sections: NavigationSection[]
-  themes: [ThemeKey, Dispatch<ThemeKey>]
   appStats?: { label: string; value: string }[]
-  advancedMode?: [boolean, Dispatch<boolean>]
-  locale: Locale
   networkName: string
 }
